pyproject.toml
[project]
name = "rest-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "knyaz7", email = "mr.vladimir.naumov@mail.ru" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "dependency-injector>=4.48.2",
    "fastapi>=0.120.3",
    "fastapi-cache2[redis]>=0.2.2",
    "pydantic-settings>=2.11.0",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[project.scripts]
rest-api-test = "rest_api_test.main:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]

[tool.ruff]
src = ["src"]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

skip-magic-trailing-comma = true

[tool.ruff.lint.isort]
split-on-trailing-comma = false
